refactor: view 와 controller 로직 분리
* view에서도 input, output 로직 분리
* output에서 RacingGame 객체 의존 제거
